[Move]InventoryUtilに移動できるメソッドを移動
